pundit auth finished

+++ app/controllers/wikis_controller.rb
@@ -1,6 +1,6 @@
 class WikisController < ApplicationController
   def index
-    @wikis = Wiki.all
+    @wikis = policy_scope(Wiki)
   end
 
   def show
@@ -9,11 +9,11 @@ class WikisController < ApplicationController
 
   def new
     @wiki = Wiki.new
-    
+
   end
 
   def create
-    @wiki = Wiki.new(wiki_params)
+    @wiki = current_user.wikis.build(wiki_params)
 
     if @wiki.save
       flash[:notice] = "Wiki was saved."
@@ -30,6 +30,7 @@ class WikisController < ApplicationController
 
   def update
     @wiki = Wiki.find(params[:id])
+    authorize @wiki
     if @wiki.update_attributes(wiki_params)
       flash[:notice] = "Wiki was updated."
       redirect_to @wiki
@@ -41,18 +42,19 @@ class WikisController < ApplicationController
 
   def destroy
       @wiki = Wiki.find(params[:id])
-        if @wiki.destroy
-          flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
-          redirect_to @wiki
-        else
-          flash[:error] = "There was an error deleting the wiki."
-          render :show
-        end
+      authorize @wiki
+      if @wiki.destroy
+        flash[:notice] = "\"#{@wiki.title}\" was deleted successfully."
+        redirect_to @wiki
+      else
+        flash[:error] = "There was an error deleting the wiki."
+        render :show
+      end
    end
 
   private
 
   def wiki_params
-    params.require(:wiki).permit(:title, :body)
+    params.require(:wiki).permit([:title, :body])
   end
 end
